Fix login functionality and add authentication hook
